More changes create avatar menu
