+++ packages/client-core/i18n/en/user.json
@@ -1,7 +1,8 @@
 {
   "common": {
     "delete": "Delete",
-    "edit": "Edit"
+    "edit": "Edit",
+    "upload": "Upload"
   },
   "auth": {
     "confirmEmail": {
@@ -372,7 +373,6 @@
     "remove": "Remove avatar",
     "uploadFiles": "Upload files",
     "useUrlInstead": "Use url instead",
-    "avatar": "Avatar",
     "cancel": "Cancel",
     "createAvatar": "Create Avatar",
     "delete": "Delete",
@@ -391,12 +391,16 @@
     "upload-success-msg": "Avatar Uploaded Successfully.",
     "remove-success-msg": "Avatar Uploaded Successfully.",
     "warning-msg": "Avatar resource is empty, have you synced avatars to your static file storage?",
-    "loadingRPM": "Loading Ready Player Me",
+    "loadingRPM": "Loading ReadyPlayer.Me",
     "downloading": "Downloading Avatar",
     "loadingPreview": "Loading Avatar Preview",
     "uploading": "Uploading Avatar",
     "searchAvatar": "Search Avatar",
-    "noAvatars": "No avatars to display"
+    "noAvatars": "No avatars to display",
+    "avatar": "Avatar",
+    "thumbnail": "Thumbnail",
+    "enterAvatarUrl": "Enter Avatar URL",
+    "enterThumbnailUrl": "Enter Thumbnail URL"
   },
   "personMenu": {
     "inviteToParty": "Invite To Party",
