+++ packages/client-core/src/common/components/Avatar/index.module.scss
@@ -12,35 +12,15 @@
     border-radius: 10px;
     border-radius: 50%;
   }
-
-  .avatarBtn {
-    position: absolute;
-    width: 30px;
-    height: 30px;
-    bottom: -10px;
-    right: -10px;
-    margin: 0;
-    min-width: 30px;
-    border-radius: 50%;
-    color: var(--textColor);
-    background: var(--iconButtonBackground);
-    box-shadow: 2px 2px 10px gray;
-    transition: all 0.15s cubic-bezier(0.18, 0.89, 0.32, 1.28);
-
-    &:hover,
-    &:focus {
-      width: 40px;
-      height: 40px;
-      bottom: -15px;
-      right: -15px;
-      background: var(--iconButtonSelectedBackground);
-      opacity: 0.7;
-    }
-  }
 }
 
 .avatarThumbnail {
   position: relative;
+}
+
+.avatarName {
+  overflow: hidden;
+  overflow-wrap: break-word;
 }
 
 .avatarRectangle {
