+++ packages/client-core/src/common/components/InputFile/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+const InputFile = React.forwardRef(
+  (
+    props: React.DetailedHTMLProps<React.InputHTMLAttributes<HTMLInputElement>, HTMLInputElement>,
+    ref: React.LegacyRef<HTMLInputElement> | undefined
+  ) => {
+    return <input ref={ref} style={{ display: 'none' }} type="file" {...props} />
+  }
+)
+
+export default InputFile
